Unattended connection (#28)
This PR introduces a very basic configuration format where, by default, a connection made to the PIM would be configured not through code but through a configuration file injected into the environment.

The client code is then as simple as:

```py
client = pypim.connect()
client.definitions(product_name="hello-world", product_version="231")
```

This is intended to support cloud scenarios, where containers would be created with configuration files injected into them, without modifying the client code. This does not close the door to application managing themselves the gRPC channel though.

The configuration file is a json file, in order to keep parsing it as simple as possible, even in other languages. The "switch" to enable the interaction with PIM is by setting the environment variable `ANSYS_PLATFORM_INSTANCEMANAGEMENT_CONFIG`. 

The configuration file format is:

```json
{
    "version": 1,
    "pim": {
        "uri": "dns:127.0.0.1:5000",
        "headers": {
            "some": "header"
        },
        "tls": false
    }
}
```