Replace raw mock creations with create_autospec (#53)
When using directly MagicMock, the mock does not know about the signature of the mocked function. It means that unless the assertion is done with "assert_called_with", then the code may be passing invalid arguments, undetected by the mock.

Using create_autospec fixes this, making the mock have the same API as the original function.

create_autospec have some caveats when used for mocking whole class, but we are only mocking functions.